# Optimise JSON marshalling for sparse histograms