--- a/model/value.go
+++ b/model/value.go
@@ -100,37 +100,19 @@ func (s Sample) MarshalJSON() ([]byte, error) {
 	return json.Marshal(&v)
 }
 
-type sampleHistogramPairPtr struct {
-	Timestamp Time
-	Histogram *SampleHistogram
-}
-
-func (s *sampleHistogramPairPtr) UnmarshalJSON(buf []byte) error {
-	tmp := []interface{}{&s.Timestamp, &s.Histogram}
-	wantLen := len(tmp)
-	if err := json.Unmarshal(buf, &tmp); err != nil {
-		return err
-	}
-	if gotLen := len(tmp); gotLen != wantLen {
-		return fmt.Errorf("wrong number of fields: %d != %d", gotLen, wantLen)
-	}
-	return nil
-}
-
 // UnmarshalJSON implements json.Unmarshaler.
-// TODO: simplify and remove the need for both sampleHistogramPairPtr and SampleHistogramPair
 func (s *Sample) UnmarshalJSON(b []byte) error {
 	v := struct {
-		Metric    Metric                 `json:"metric"`
-		Value     SamplePair             `json:"value"`
-		Histogram sampleHistogramPairPtr `json:"histogram"`
+		Metric    Metric              `json:"metric"`
+		Value     SamplePair          `json:"value"`
+		Histogram SampleHistogramPair `json:"histogram"`
 	}{
 		Metric: s.Metric,
 		Value: SamplePair{
 			Timestamp: s.Timestamp,
 			Value:     s.Value,
 		},
-		Histogram: sampleHistogramPairPtr{
+		Histogram: SampleHistogramPair{
 			Timestamp: s.Timestamp,
 			Histogram: s.Histogram,
 		},