Follow up to https://github.com/prometheus/common/pull/417 according to suggestion in https://github.com/prometheus/common/pull/417#discussion_r1050576656

Benchmarking results:
```
name                                         old time/op    new time/op    delta
JSONMarshallingSamplePairMatrix-16             6.96µs ± 8%    3.44µs ± 4%  -50.52%  (p=0.000 n=20+19)
JSONMarshallingSampleHistogramPairMatrix-16    69.2µs ± 4%    18.9µs ± 4%  -72.65%  (p=0.000 n=20+19)

name                                         old alloc/op   new alloc/op   delta
JSONMarshallingSamplePairMatrix-16             1.63kB ± 0%    0.84kB ± 0%  -48.31%  (p=0.000 n=20+20)
JSONMarshallingSampleHistogramPairMatrix-16    20.3kB ± 0%     4.6kB ± 0%  -77.17%  (p=0.000 n=20+20)

name                                         old allocs/op  new allocs/op  delta
JSONMarshallingSamplePairMatrix-16               74.0 ± 0%      22.0 ± 0%  -70.27%  (p=0.000 n=20+20)
JSONMarshallingSampleHistogramPairMatrix-16       658 ± 0%        22 ± 0%  -96.66%  (p=0.000 n=20+20)
```