@zenador 2023-02-01
Thanks for the review! Updated benchmark results:
```
name                                         old time/op    new time/op    delta
JSONMarshallingSamplePairMatrix-16             6.66µs ± 4%    3.36µs ± 1%  -49.58%  (p=0.000 n=20+19)
JSONMarshallingSampleHistogramPairMatrix-16    68.2µs ± 4%    18.2µs ± 1%  -73.27%  (p=0.000 n=19+20)

name                                         old alloc/op   new alloc/op   delta
JSONMarshallingSamplePairMatrix-16             1.63kB ± 0%    0.84kB ± 0%  -48.31%  (p=0.000 n=18+20)
JSONMarshallingSampleHistogramPairMatrix-16    20.3kB ± 0%     4.6kB ± 0%  -77.17%  (p=0.000 n=20+20)

name                                         old allocs/op  new allocs/op  delta
JSONMarshallingSamplePairMatrix-16               74.0 ± 0%      22.0 ± 0%  -70.27%  (p=0.000 n=20+20)
JSONMarshallingSampleHistogramPairMatrix-16       658 ± 0%        22 ± 0%  -96.66%  (p=0.000 n=20+20)
```